+++ packages/ts-components/src/components/in-article-puff/InArticlePuff.stories.tsx
@@ -6,13 +6,7 @@ import { FetchProvider } from '../../helpers/fetch/FetchProvider';
 import { InArticlePuff } from './InArticlePuff';
 import { TrackingContextProvider } from '../../helpers/tracking/TrackingContextProvider';
 
-import { action } from '@storybook/addon-actions';
-
-const analyticsStream = (event: any) => {
-  // tslint:disable-next-line:no-console
-  console.log('analytics-action', event);
-  action('analytics-action')(event);
-};
+import analyticsStream from '../../fixtures/analytics-actions/analytics-actions';
 
 const showcase = {
   children: [
@@ -63,9 +57,18 @@ const showcase = {
       ),
       name: 'In Article Puff - No Image',
       type: 'story'
+    },
+    {
+      component: () => (
+        <FetchProvider url="https://gobble.timesdev.tools/deck/api/deck-post-action/41547">
+          <InArticlePuff sectionColour="#184e13" sanitiseHtml />
+        </FetchProvider>
+      ),
+      name: 'In Article Puff - Sanitised',
+      type: 'story'
     }
   ],
-  name: 'Typescript Component/In Article Puff'
+  name: 'Typescript Component/In Article/In Article Puff'
 };
 
 // @ts-ignore