+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.5](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.4...@times-components/article-extras@0.7.5) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.7.4](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.3...@times-components/article-extras@0.7.4) (2021-05-27)
 
 **Note:** Version bump only for package @times-components/article-extras